+++ app/components/creators/shared/buttons/index.js
@@ -1,3 +1,5 @@
 export { default as MediumButton } from './MediumButton';
 export { default as JoinNowButton } from './JoinNowButton';
 export { default as ProjectButton } from './ProjectButton';
+export { default as BlueButton } from './BlueButton';
+export { default as GrayButton } from './GrayButton';
