+++ app/components/creators/shared/modals/Join/index.js
@@ -7,7 +7,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { Modal, Fade } from '@material-ui/core';
+import { Fade } from '@material-ui/core';
 import LogoCaps from 'images/logo.svg';
 import { Body, Body18, Body13 } from '../../typography';
 import { MediumButton } from '../../buttons';
@@ -15,23 +15,7 @@ import Heads from '../../Heads';
 import FacebookIcon from 'images/icons/facebook.svg';
 import TwitterIcon from 'images/icons/twitter.svg';
 import EmailIcon from 'images/icons/email.svg';
-
-const Wrapper = styled.div`
-  max-width: 36rem;
-  width: 100%;
-  padding: 3rem;
-  background-color: #fff;
-  border-radius: 0.5rem;
-  border: none;
-  text-align: center;
-`;
-const OverlayModal = styled(Modal)`
-  && {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-`;
+import { BodyWrapper, OverlayModal } from '../shared';
 
 const Logo = styled.img`
   height: 97px;
@@ -88,12 +72,12 @@ function Join({ open, handleClose }) {
       aria-describedby="simple-modal-description"
     >
       <Fade in={open}>
-        <Wrapper>
+        <BodyWrapper>
           <Logo src={LogoCaps} alt="logo" />
           <Title>Fixing politics for Good!</Title>
           <Blurb>
-            Join our community of creators <br /> working together for the Good of
-            all.
+            Join our community of creators <br /> working together for the Good
+            of all.
           </Blurb>
           <Heads />
           <JoinButtonWrapper>
@@ -107,8 +91,10 @@ function Join({ open, handleClose }) {
               <SocialIcon src={EmailIcon} /> Join With Email
             </JoinWithButton>
           </JoinButtonWrapper>
-          <FooterMessage>We will never post to any of your accounts without your permission.</FooterMessage>
-        </Wrapper>
+          <FooterMessage>
+            We will never post to any of your accounts without your permission.
+          </FooterMessage>
+        </BodyWrapper>
       </Fade>
     </OverlayModal>
   );
