+++ app/components/creators/shared/modals/index.js
@@ -1,3 +1,3 @@
 export { default as Join } from './Join';
-export { default as TouchProject } from './TouchProject';
+export { default as ProjectProposal } from './ProjectProposal';
 export { default as ListProject } from './ListProject';
