+++ app/components/creators/shared/modals/shared/BodyWrapper.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const BodyWrapper = styled.div`
+  max-width: 40rem;
+  width: 100%;
+  padding: 3rem;
+  background-color: #fff;
+  border-radius: 0.5rem;
+  border: none;
+  text-align: center;
+`;
+
+export default BodyWrapper;
