+++ app/components/creators/shared/modals/shared/FooterWrapper.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const FooterWrapper = styled.div`
+    display: flex;
+    align-items: center;
+    margin-top: 2.5rem;
+`;
+
+export default FooterWrapper;
