+++ app/components/creators/shared/modals/shared/OverlayModal.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+import { Modal } from '@material-ui/core';
+
+const OverlayModal = styled(Modal)`
+  && {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+`;
+
+export default OverlayModal;
