+++ app/components/creators/shared/modals/shared/ProjectFormControl.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import { FormControl } from '@material-ui/core';
+
+const ProjectFormControl = styled(FormControl)`
+  && {
+    margin-bottom: 2.5rem;
+  }
+`;
+
+export default ProjectFormControl;
+
