+++ app/components/creators/shared/modals/shared/ProjectFormControlHeader.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const ProjectFormControlHeader = styled.div`
+  display: flex;
+`;
+
+export default ProjectFormControlHeader;
