+++ app/components/creators/shared/modals/shared/ProjectFormLabel.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import { Body18 } from '../../typography';
+
+const ProjectFormLabel = styled(Body18)`
+  color: black;
+  text-transform: uppercase;
+  font-weight: 600;
+  margin-right: 0.5rem;
+`;
+
+export default ProjectFormLabel;
