+++ app/components/creators/shared/modals/shared/Title.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import { Body } from '../../typography';
+
+const Title = styled(Body)`
+  color: #000;
+  line-height: 130%;
+  text-transform: none;
+  margin-top: 0;
+  margin-bottom: 2.5rem;
+  text-align: left;
+`;
+
+export default Title;
