+++ app/components/creators/shared/modals/shared/index.js
@@ -1,0 +1,9 @@
+export { default as BodyWrapper } from './BodyWrapper';
+export { default as FooterWrapper } from './FooterWrapper';
+export { default as OverlayModal } from './OverlayModal';
+export { default as Title } from './Title';
+export { default as FormText } from './FormText';
+export { default as ProjectFormLabel } from './ProjectFormLabel';
+export { default as ProjectFormControl } from './ProjectFormControl';
+export { default as ProjectFormHelperText } from './ProjectFormHelperText';
+export { default as ProjectFormControlHeader } from './ProjectFormControlHeader';
